pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ams-compose"
version = "0.0.0"
description = "Dependency management tool for analog/mixed-signal IC design repositories"
authors = [
    {name = "ams-compose contributors"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Software Distribution",
]
keywords = ["analog", "mixed-signal", "ams", "ic", "design", "dependency", "management", "eda"]

dependencies = [
    "GitPython>=3.1.40",
    "pydantic>=2.0.0",
    "click>=8.1.0",
    "PyYAML>=6.0.0",
    "pathspec>=0.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=2.0.0",
    "sphinx-autodoc-typehints>=1.24.0",
    "myst-parser>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/Jianxun/ams-compose"
Repository = "https://github.com/Jianxun/ams-compose"
Issues = "https://github.com/Jianxun/ams-compose/issues"

[project.scripts]
ams-compose = "ams_compose.cli.main:main"

[tool.setuptools.packages.find]
include = ["ams_compose*"]

[tool.setuptools.package-data]
ams_compose = ["schema.md"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=ams_compose --cov-report=term-missing"

[tool.sphinx]
source-dir = "docs"
build-dir = "docs/_build"